test cl now runs on Python 3.5
Also added per-case grading so we won't count multiple issues in one
case as multiple failures but just a failing of that test case.

The warnings and error checker will use regex to find matches and will
also log the warnings and errors we got to the result text file so the
student can see this output when we grade them.