Revert to not automatically pulling late submissions
To download a late submission you must explicitly request the student's
canvas ID and a warning will be printed that the submission is late.
When downloading regularly a warning is printed if we're pulling a
submission that they have newer (but late) submissions on. The allowLate
flag is used by the autograder to always accept late submissions so late
students can still get feedback.